Rename ingestionLogicV2 to just ingestionLogic
