+++ frontend/src/scenes/ingestion/panels/PlatformPanel.tsx
@@ -1,5 +1,5 @@
 import { useActions } from 'kea'
-import { ingestionLogicV2 } from '../ingestionLogicV2'
+import { ingestionLogic } from '../ingestionLogic'
 import { THIRD_PARTY, platforms } from '../constants'
 import { LemonButton } from 'lib/lemon-ui/LemonButton'
 import './Panels.scss'
@@ -7,7 +7,7 @@ import { LemonDivider } from 'lib/lemon-ui/LemonDivider'
 import { IngestionInviteMembersButton } from '../IngestionInviteMembersButton'
 
 export function PlatformPanel(): JSX.Element {
-    const { next } = useActions(ingestionLogicV2)
+    const { next } = useActions(ingestionLogic)
 
     return (
         <div>
