+++ frontend/src/scenes/ingestion/panels/TeamInvitedPanel.tsx
@@ -1,5 +1,5 @@
 import { useActions } from 'kea'
-import { ingestionLogicV2 } from 'scenes/ingestion/ingestionLogicV2'
+import { ingestionLogic } from 'scenes/ingestion/ingestionLogic'
 import { LemonButton } from 'lib/lemon-ui/LemonButton'
 import './Panels.scss'
 import { LemonDivider } from 'lib/lemon-ui/LemonDivider'
@@ -8,7 +8,7 @@ import { eventUsageLogic } from 'lib/utils/eventUsageLogic'
 import { DemoProjectButton } from './PanelComponents'
 
 export function TeamInvitedPanel(): JSX.Element {
-    const { completeOnboarding } = useActions(ingestionLogicV2)
+    const { completeOnboarding } = useActions(ingestionLogic)
     const { reportIngestionContinueWithoutVerifying } = useActions(eventUsageLogic)
 
     return (
